test: split tests into smaller files
