+++ catalog/internal/operations/memory_store_test.go
@@ -1,4 +1,4 @@
-package catalog
+package operations
 
 import (
 	"errors"
@@ -13,13 +13,13 @@ func newTestOperation(name, plugin string, createdAt time.Time) Operation {
 	return Operation{
 		Name:      name,
 		Plugin:    plugin,
-		State:     OperationStatePending,
+		State:     StatePending,
 		CreatedAt: createdAt,
 	}
 }
 
-func TestMemoryOperationStoreCreateAndGet(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreCreateAndGet(t *testing.T) {
+	store := NewMemoryStore()
 	op := newTestOperation("operations/plugin-run-1", "mlflow", time.Now())
 
 	require.NoError(t, store.Create(op))
@@ -29,110 +29,110 @@ func TestMemoryOperationStoreCreateAndGet(t *testing.T) {
 	assert.Equal(t, op, got)
 }
 
-func TestMemoryOperationStoreGetNotFound(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreGetNotFound(t *testing.T) {
+	store := NewMemoryStore()
 
 	_, err := store.Get("operations/plugin-run-missing")
 	require.Error(t, err)
-	assert.True(t, errors.Is(err, ErrOperationNotFound))
+	assert.True(t, errors.Is(err, ErrNotFound))
 }
 
-func TestMemoryOperationStoreCreateDuplicateName(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreCreateDuplicateName(t *testing.T) {
+	store := NewMemoryStore()
 	op := newTestOperation("operations/plugin-run-1", "mlflow", time.Now())
 
 	require.NoError(t, store.Create(op))
 	err := store.Create(op)
 	require.Error(t, err)
-	assert.True(t, errors.Is(err, ErrOperationAlreadyExists))
+	assert.True(t, errors.Is(err, ErrAlreadyExists))
 }
 
-func TestMemoryOperationStoreUpdateStateSetsRunningStartTime(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreUpdateStateSetsRunningStartTime(t *testing.T) {
+	store := NewMemoryStore()
 	op := newTestOperation("operations/plugin-run-1", "mlflow", time.Now())
 	require.NoError(t, store.Create(op))
 
-	require.NoError(t, store.UpdateState(op.Name, OperationStateRunning, nil, 0, 0))
+	require.NoError(t, store.UpdateState(op.Name, StateRunning, nil, 0, 0))
 
 	got, err := store.Get(op.Name)
 	require.NoError(t, err)
-	assert.Equal(t, OperationStateRunning, got.State)
+	assert.Equal(t, StateRunning, got.State)
 	assert.False(t, got.StartTime.IsZero(), "start time should be set on RUNNING")
 	assert.Nil(t, got.EndTime, "end time should be nil while running")
 }
 
-func TestMemoryOperationStoreUpdateStateSucceededSetsResult(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreUpdateStateSucceededSetsResult(t *testing.T) {
+	store := NewMemoryStore()
 	op := newTestOperation("operations/plugin-run-1", "mlflow", time.Now())
 	require.NoError(t, store.Create(op))
 
-	require.NoError(t, store.UpdateState(op.Name, OperationStateRunning, nil, 0, 0))
-	require.NoError(t, store.UpdateState(op.Name, OperationStateSucceeded, nil, 3, 5))
+	require.NoError(t, store.UpdateState(op.Name, StateRunning, nil, 0, 0))
+	require.NoError(t, store.UpdateState(op.Name, StateSucceeded, nil, 3, 5))
 
 	got, err := store.Get(op.Name)
 	require.NoError(t, err)
-	assert.Equal(t, OperationStateSucceeded, got.State)
+	assert.Equal(t, StateSucceeded, got.State)
 	assert.NotNil(t, got.EndTime, "end time should be set on SUCCEEDED")
 	assert.Equal(t, 3, got.NodesUpserted)
 	assert.Equal(t, 5, got.RelationsUpserted)
 	assert.Nil(t, got.Error)
 }
 
-func TestMemoryOperationStoreUpdateStateFailedSetsError(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreUpdateStateFailedSetsError(t *testing.T) {
+	store := NewMemoryStore()
 	op := newTestOperation("operations/plugin-run-1", "mlflow", time.Now())
 	require.NoError(t, store.Create(op))
 
 	statusErr := &StatusError{Message: "boom"}
-	require.NoError(t, store.UpdateState(op.Name, OperationStateFailed, statusErr, 0, 0))
+	require.NoError(t, store.UpdateState(op.Name, StateFailed, statusErr, 0, 0))
 
 	got, err := store.Get(op.Name)
 	require.NoError(t, err)
-	assert.Equal(t, OperationStateFailed, got.State)
+	assert.Equal(t, StateFailed, got.State)
 	assert.NotNil(t, got.EndTime)
 	assert.Equal(t, statusErr, got.Error)
 }
 
-func TestMemoryOperationStoreUpdateStateNotFound(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreUpdateStateNotFound(t *testing.T) {
+	store := NewMemoryStore()
 
-	err := store.UpdateState("operations/missing", OperationStateRunning, nil, 0, 0)
+	err := store.UpdateState("operations/missing", StateRunning, nil, 0, 0)
 	require.Error(t, err)
-	assert.True(t, errors.Is(err, ErrOperationNotFound))
+	assert.True(t, errors.Is(err, ErrNotFound))
 }
 
-func TestMemoryOperationStoreListFiltersByPlugin(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreListFiltersByPlugin(t *testing.T) {
+	store := NewMemoryStore()
 	now := time.Now()
 
 	require.NoError(t, store.Create(newTestOperation("operations/plugin-run-1", "mlflow", now)))
 	require.NoError(t, store.Create(newTestOperation("operations/plugin-run-2", "litellm", now)))
 
-	ops, err := store.List(OperationFilter{Plugin: "mlflow"})
+	ops, err := store.List(Filter{Plugin: "mlflow"})
 	require.NoError(t, err)
 	require.Len(t, ops, 1)
 	assert.Equal(t, "mlflow", ops[0].Plugin)
 }
 
-func TestMemoryOperationStoreListFiltersByState(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreListFiltersByState(t *testing.T) {
+	store := NewMemoryStore()
 	now := time.Now()
 
 	op1 := newTestOperation("operations/plugin-run-1", "mlflow", now)
 	require.NoError(t, store.Create(op1))
-	require.NoError(t, store.UpdateState(op1.Name, OperationStateSucceeded, nil, 1, 1))
+	require.NoError(t, store.UpdateState(op1.Name, StateSucceeded, nil, 1, 1))
 
 	op2 := newTestOperation("operations/plugin-run-2", "litellm", now)
 	require.NoError(t, store.Create(op2))
 
-	ops, err := store.List(OperationFilter{State: OperationStateSucceeded})
+	ops, err := store.List(Filter{State: StateSucceeded})
 	require.NoError(t, err)
 	require.Len(t, ops, 1)
 	assert.Equal(t, "mlflow", ops[0].Plugin)
 }
 
-func TestMemoryOperationStoreListOrdersMostRecentFirst(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreListOrdersMostRecentFirst(t *testing.T) {
+	store := NewMemoryStore()
 	base := time.Now()
 
 	oldOp := newTestOperation("operations/plugin-run-old", "mlflow", base.Add(-2*time.Hour))
@@ -141,19 +141,19 @@ func TestMemoryOperationStoreListOrdersMostRecentFirst(t *testing.T) {
 	require.NoError(t, store.Create(oldOp))
 	require.NoError(t, store.Create(newOp))
 
-	ops, err := store.List(OperationFilter{})
+	ops, err := store.List(Filter{})
 	require.NoError(t, err)
 	require.Len(t, ops, 2)
 	assert.Equal(t, newOp.Name, ops[0].Name)
 	assert.Equal(t, oldOp.Name, ops[1].Name)
 }
 
-func TestMemoryOperationStoreEvictsOldestPerPlugin(t *testing.T) {
-	store := NewMemoryOperationStore()
+func TestMemoryStoreEvictsOldestPerPlugin(t *testing.T) {
+	store := NewMemoryStore()
 	base := time.Now()
 
-	// Insert more than the per-plugin cap (5) for the same plugin.
-	for i := 0; i < maxOperationsPerPlugin+1; i++ {
+	// Insert more than the per-plugin cap for the same plugin.
+	for i := range maxPerPlugin + 1 {
 		op := newTestOperation(
 			"operations/plugin-run-"+string(rune('a'+i)),
 			"mlflow",
@@ -162,11 +162,11 @@ func TestMemoryOperationStoreEvictsOldestPerPlugin(t *testing.T) {
 		require.NoError(t, store.Create(op))
 	}
 
-	ops, err := store.List(OperationFilter{Plugin: "mlflow"})
+	ops, err := store.List(Filter{Plugin: "mlflow"})
 	require.NoError(t, err)
-	require.Len(t, ops, maxOperationsPerPlugin)
+	require.Len(t, ops, maxPerPlugin)
 
 	// The oldest operation ("operations/plugin-run-a") must have been evicted.
 	_, err = store.Get("operations/plugin-run-a")
-	assert.True(t, errors.Is(err, ErrOperationNotFound))
+	assert.True(t, errors.Is(err, ErrNotFound))
 }
